chore: rename sample app name
