+++ samples/next-app-with-fetch/.eslintrc.json
