+++ samples/next-app-with-fetch/app/favicon.ico
