+++ samples/next-app-with-fetch/app/gen/models/cat.ts
