+++ samples/next-app-with-fetch/app/gen/models/catType.ts
