+++ samples/next-app-with-fetch/app/gen/models/createPetsBodyItem.ts
