+++ samples/next-app-with-fetch/app/gen/models/dachshund.ts
