+++ samples/next-app-with-fetch/app/gen/models/dachshundBreed.ts
