+++ samples/next-app-with-fetch/app/gen/models/dog.ts
