+++ samples/next-app-with-fetch/app/gen/models/dogType.ts
