+++ samples/next-app-with-fetch/app/gen/models/error.ts
