+++ samples/next-app-with-fetch/app/gen/models/index.ts
