+++ samples/next-app-with-fetch/app/gen/models/labradoodle.ts
