+++ samples/next-app-with-fetch/app/gen/models/labradoodleBreed.ts
