+++ samples/next-app-with-fetch/app/gen/models/listPetsParams.ts
