+++ samples/next-app-with-fetch/app/gen/models/pet.ts
