+++ samples/next-app-with-fetch/app/gen/models/petCallingCode.ts
