+++ samples/next-app-with-fetch/app/gen/models/petCountry.ts
