+++ samples/next-app-with-fetch/app/gen/models/pets.ts
