+++ samples/next-app-with-fetch/app/globals.css
