+++ samples/next-app-with-fetch/app/layout.tsx
