+++ samples/next-app-with-fetch/app/pets.tsx
