+++ samples/next-app-with-fetch/next.config.mjs
