+++ samples/next-app-with-fetch/orval.config.ts
