+++ samples/next-app-with-fetch/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "next-fetch",
+  "name": "next-app-with-fetch",
   "version": "0.1.0",
   "private": true,
   "scripts": {
