+++ samples/next-app-with-fetch/petstore.yaml
