+++ samples/next-app-with-fetch/tailwind.config.ts
