+++ samples/next-app-with-fetch/tsconfig.json
